+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# v0.1.0 (Jan 8, 2023)
+# v1.1.1 (Jan 8, 2023)
+
+1.  Allow watching file even if the file didn't exist.
+
+# v1.1.0 (Jan 8, 2023)
+
+1.  Allow converting a value to time.Duration.
+
+# v1.0.0 (Jan 4, 2023)
 
 1.  Support json, ini format.
 2.  Reload config when the file changes.